requirements.txt
alembic==1.6.5
appdirs==1.4.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
idna==2.10
importlib-metadata==4.4.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
peewee==3.14.4
psycopg2==2.9.1
pydantic==1.8.1
PyJWT==2.1.0
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
regex==2021.4.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.11
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.5
uvicorn==0.13.4
zipp==3.4.1
